Renamed command to sam remote invoke
