+++ samcli/cli/command.py
@@ -30,7 +30,7 @@ _SAM_CLI_COMMAND_PACKAGES = [
     "samcli.commands.pipeline.pipeline",
     "samcli.commands.list.list",
     "samcli.commands.docs",
-    # "samcli.commands.cloud.cloud",
+    # "samcli.commands.remote.remote",
     # We intentionally do not expose the `bootstrap` command for now. We might open it up later
     # "samcli.commands.bootstrap",
 ]
