+++ tests/unit/commands/remote/test_remote_invoke_context.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 from unittest.mock import Mock, patch
 from uuid import uuid4
 
-from samcli.commands.cloud.exceptions import (
+from samcli.commands.remote.exceptions import (
     InvalidRemoteInvokeParameters,
     AmbiguousResourceForRemoteInvoke,
     NoResourceFoundForRemoteInvoke,
@@ -10,7 +10,7 @@ from samcli.commands.cloud.exceptions import (
     NoExecutorFoundForRemoteInvoke,
     InvalidStackNameProvidedForRemoteInvoke,
 )
-from samcli.commands.cloud.remote_invoke_context import RemoteInvokeContext, SUPPORTED_SERVICES
+from samcli.commands.remote.remote_invoke_context import RemoteInvokeContext, SUPPORTED_SERVICES
 from samcli.lib.utils.cloudformation import CloudFormationResourceSummary
 
 
@@ -33,7 +33,7 @@ class TestRemoteInvokeContext(TestCase):
             with self._get_remote_invoke_context():
                 pass
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summaries")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summaries")
     def test_invalid_stack_name_with_no_resource_should_fail(self, patched_resource_summaries):
         self.resource_id = None
         patched_resource_summaries.side_effect = InvalidStackNameProvidedForRemoteInvoke("Invalid stack-name")
@@ -41,7 +41,7 @@ class TestRemoteInvokeContext(TestCase):
             with self._get_remote_invoke_context():
                 pass
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summaries")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summaries")
     def test_only_stack_name_with_no_resource_should_fail(self, patched_resource_summaries):
         self.resource_id = None
         patched_resource_summaries.return_value = {}
@@ -49,7 +49,7 @@ class TestRemoteInvokeContext(TestCase):
             with self._get_remote_invoke_context():
                 pass
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summaries")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summaries")
     def test_only_stack_name_with_multiple_resource_should_fail(self, patched_resource_summaries):
         self.resource_id = None
         patched_resource_summaries.return_value = {"resource1": Mock(), "resource2": Mock()}
@@ -57,7 +57,7 @@ class TestRemoteInvokeContext(TestCase):
             with self._get_remote_invoke_context():
                 pass
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summaries")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summaries")
     def test_only_stack_name_with_single_resource_should_be_valid(self, patched_resource_summaries):
         self.resource_id = None
         resource_summary = Mock(logical_resource_id=self.resource_id)
@@ -84,7 +84,7 @@ class TestRemoteInvokeContext(TestCase):
                 ),
             )
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summary_from_physical_id")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summary_from_physical_id")
     def test_only_resource_id_as_invalid_physical_id_should_fail(self, patched_resource_summary_from_physical_id):
         self.stack_name = None
         patched_resource_summary_from_physical_id.return_value = None
@@ -92,14 +92,14 @@ class TestRemoteInvokeContext(TestCase):
             with self._get_remote_invoke_context():
                 pass
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summary")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summary")
     def test_if_no_resource_found_with_given_stack_and_resource_id_should_fail(self, patched_get_resource_summary):
         patched_get_resource_summary.return_value = None
         with self.assertRaises(AmbiguousResourceForRemoteInvoke):
             with self._get_remote_invoke_context() as remote_invoke_context:
                 remote_invoke_context.run(Mock())
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summary_from_physical_id")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summary_from_physical_id")
     def test_only_resource_id_as_valid_physical_id_should_be_valid(self, patched_resource_summary_from_physical_id):
         self.stack_name = None
         resource_summary = Mock()
@@ -107,22 +107,22 @@ class TestRemoteInvokeContext(TestCase):
         with self._get_remote_invoke_context() as remote_invoke_context:
             self.assertEqual(remote_invoke_context._resource_summary, resource_summary)
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summary")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summary")
     def test_running_without_resource_summary_should_raise_exception(self, patched_get_resource_summary):
         patched_get_resource_summary.return_value = None
         with self._get_remote_invoke_context() as remote_invoke_context:
             with self.assertRaises(AmbiguousResourceForRemoteInvoke):
                 remote_invoke_context.run(Mock())
 
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summary")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summary")
     def test_running_with_unsupported_resource_should_raise_exception(self, patched_get_resource_summary):
         patched_get_resource_summary.return_value = Mock(resource_type="UnSupportedResource")
         with self._get_remote_invoke_context() as remote_invoke_context:
             with self.assertRaises(NoExecutorFoundForRemoteInvoke):
                 remote_invoke_context.run(Mock())
 
-    @patch("samcli.commands.cloud.remote_invoke_context.RemoteInvokeExecutorFactory")
-    @patch("samcli.commands.cloud.remote_invoke_context.get_resource_summary")
+    @patch("samcli.commands.remote.remote_invoke_context.RemoteInvokeExecutorFactory")
+    @patch("samcli.commands.remote.remote_invoke_context.get_resource_summary")
     def test_running_should_execute_remote_invoke_executor_instance(
         self, patched_get_resource_summary, patched_remote_invoke_executor_factory
     ):
